player_valuations/pyproject.toml
[project]
name = "player-valuations"
version = "0.1.0"
description = "Valuation Pathway Simulator: quantify how exposure regime transitions affect player valuation distributions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.0",
    "pytest>=7.0.0",
]

[project.scripts]
simulate = "valuation_pathways.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["valuation_pathways"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
